Remove code duplicacy
