+++ item_service/item/src/item_service_api/delivery_option.rs
@@ -1,0 +1,8 @@
+use crate::item_service_api::location::Location;
+
+pub enum DeliveryOption {
+    Null,
+    PickUp,
+    Deliver {name: String, price: int, location: Location},
+    ByNegotiation {location: Location},
+}
