+++ item_service/item/src/item_service_api/item.rs
@@ -1,4 +1,10 @@
+use crate::item_service_api::item_data::ItemData;
+use crate::item_service_api::item_status::ItemStatus;
+use crate::item_service_api::location::Location;
+use crate::item_service_impl::controller::error::CustomError;
+use crate::item_service_impl::constants::constants::PRICE;
 use std::ptr::null;
+use crate::item_service_api::delivery_option::DeliveryOption;
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Item {
@@ -15,33 +21,48 @@ pub struct Item {
 }
 
 impl Item {
-    pub fn new(id: i32, creator: String,
+    fn new(id: i32, creator: String,
                item_data: ItemData,
                price: f32,
                status: ItemStatus,
                auction_start: Option<String>,
                auction_end: Option<String>,
                auction_winner: Option<String>) -> Item {
-        let item_status: Result<ItemStatus,CustomError> = match status {
-            ItemStatus::CREATED => Ok(status),
-            ItemStatus::AUCTION => Ok(status),
-            ItemStatus::CANCELLED => Ok(status),
-            ItemStatus::COMPLETED => Ok(status),
-            _ => Err(CustomError::InvalidInput{field:"status doesn't set"})
-        };
+
         Item {
             id,
             creator,
             item_data,
             price: if price != PRICE { price } else { PRICE },
-            status: item_status,
+            status: if status != ItemStatus::NULL { status} else {ItemStatus::NULL},
             auction_start,
             auction_end,
             auction_winner,
             location: Location{
-
+                country: None,
+                state: None,
+                city: None
             },
-            delivery_options: ()
+            delivery_options: DeliveryOption::Null,
         }
+    }
+
+    pub fn new_item(&self,id: i32, creator: String,
+           item_data: ItemData,
+           price: f32,
+           status: ItemStatus,
+           auction_start: Option<String>,
+           auction_end: Option<String>,
+           auction_winner: Option<String>) {
+
+            self.id: id;
+            self.creator: creator,
+        selfitem_data,
+        selfprice,
+        selfstatus,
+        selfauction_start,
+        selfauction_end,
+        selfauction_winner,
+            }
     }
 }
