+++ item_service/item/src/item_service_api/item_event.rs
@@ -1,0 +1,11 @@
+use crate::item_service_api::item::Item;
+use crate::item_service_impl::models::item_status::ItemStatus;
+
+pub enum ItemEvent {
+    ItemUpdated { item_id: i32, creator: String, title: String, description: String, item_status:
+    ItemStatus, currency_id: String },
+    AuctionStarted { item_id: i32, creator: String, reserve_price: f32, increment: f32, start_date:
+    String, end_date: String },
+    AuctionFinished { iyem_id: i32, item: Item},
+    AuctionCancelled {item_id: i32}
+}
