+++ item_service/item/src/item_service_api/item_status.rs
@@ -1,0 +1,8 @@
+#[derive(Debug,Clone,PartialEq,Serialize, Deserialize)]
+pub enum ItemStatus {
+    NULL,
+    CREATED,
+    AUCTION,
+    COMPLETED,
+    CANCELLED,
+}
