+++ item_service/item/src/item_service_api/item_summary.rs
@@ -1,0 +1,8 @@
+pub struct ItemSummary {
+    pub id: i32,
+    pub title: String,
+    pub currency_id: String,
+    pub reserve_price: f32,
+    pub status: ItemStatus,
+
+}
