--- item_service/item/src/item_service_impl/models/item_summary.rs
@@ -1,10 +1,0 @@
-use crate::models::item_status::ItemStatus;
-
-pub struct ItemSummary {
-    pub id: i32,
-    pub title: String,
-    pub currency_id: String,
-    pub reserve_price: f32,
-    pub status: ItemStatus,
-
-}
